+++ CHANGELOG.md
@@ -15,14 +15,24 @@ File descriptors metric now tries to use lsof | wc first (#359)
 #### Updated
 
 
+## [2.0.0-alpha.11] - 2021-12-10
+#### Added
+- Hierarchical filters (#333)
+
+
+#### Fixed
+- Disk space metrics calculation fix (#376)
+
+
+#### Updated
+- Upgrade pearl to v0.9.0
+
+
 ## [2.0.0-alpha.10] - 2021-12-02
 #### Added
 - Add bloom_filter_memory_limit to example config (#378)
 - Add traits to support pearl #123
 
-
-#### Fixed
-- Cluster tests now use mock of BobClient, some cluster and cluster config tests were updated, all tests now pass (#280)
 
 #### Updated
 - Upgrade pearl to v0.8.1